Adding helper file for edge spoke transit attachment
